requirements.txt
numpy~=1.25.2
openai~=0.28.0
pandas~=2.1.0
seaborn~=0.12.2
scikit-learn~=1.3.1
pytest
fastapi[all]~=0.103.2
python-dotenv~=1.0.0
bubble_api~=0.5.0